Restack and fix tests
